# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Projeto Integrador",
    "version": "1.0.0",
    "description": "Documentação básica das principais rotas da API."
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "paths": {
...
    "/categoria": {
      "get": {
        "summary": "Listar todas as categorias",
        "tags": ["Categorias"],
        "responses": {
          "200": {
            "description": "Lista de categorias",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Categoria" } }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar nova categoria",
        "tags": ["Categorias"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoriaInput" },
              "example": { "nome": "Roupas" }
            }
          }
        },
        "responses": {
          "201": { "description": "Categoria criada" },
          "400": { "description": "Dados inválidos" }
        }
      }
    },
    "/categoria/{id}": {
      "get": {
        "summary": "Buscar categoria por ID",
        "tags": ["Categorias"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Categoria encontrada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Categoria" } } } },
          "404": { "description": "Categoria não encontrada" }
        }
      },
      "put": {
        "summary": "Atualizar categoria",
        "tags": ["Categorias"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoriaInput" }
            }
          }
        },
        "responses": {
          "200": { "description": "Categoria atualizada" },
          "404": { "description": "Categoria não encontrada" }
        }
      },
      "delete": {
        "summary": "Deletar categoria",
        "tags": ["Categorias"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Categoria deletada" },
          "404": { "description": "Categoria não encontrada" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Categoria": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "nome": { "type": "string" },
          "criado_em": { "type": "string", "format": "date-time" },
          "atualizado_em": { "type": "string", "format": "date-time" }
        }
      },
      "CategoriaInput": {
        "type": "object",
        "properties": {
          "nome": { "type": "string" }
        },
        "required": ["nome"]
      },
...
    }
  }
}